// src/api/gen/_model/recommendationResponse.gen.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * TMT API
 * 또맛또(TMT) API
 * OpenAPI spec version: v1
 */
import type { RecommendedPlace } from "./recommendedPlace.gen";
import type { RecommendedSummary } from "./recommendedSummary.gen";

export interface RecommendationResponse {
  place: RecommendedPlace;
  summary?: RecommendedSummary | null;
}